using .h file
